提交支持客服知识库，机器人知识库，机器人测试聊天代码

+++ gemini-kbqa-admin/gemini-kbqa-admin-customerKb/src/main/java/com/gemini/admin/dao/CustomerKbDaoImpl.java
@@ -75,7 +75,7 @@ public class CustomerKbDaoImpl {
         }
     }
 
-    public void deleteKbCustomerArticle(Integer id){
+    public void deleteKbCustomerArticle(Long id){
         Integer row = kbCustomerArticleMapper.deleteById(id);
         if(row == 0){
             log.warn("删除文档失败:{}", id);
